+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Redesigned project overview page [#365](https://github.com/DigitalExcellence/dex-frontend/issues/365)
 
 ### Deprecated